refactor the genre listing page & partially SSR
