+++ src/app/[genre]/page.tsx
@@ -1,11 +1,31 @@
-"use client";
+import RenderLyricsList from "@/src/components/RenderLyricsList";
+import { WEB_BASE_URL } from "@/src/utilities/constants";
+import { capitalize, getPageGenre } from "@/src/utilities/helpers";
+import { Metadata } from "next";
 
-import Loader from "@/src/components/Loader";
-import GenreInfo from "@/src/models/GenreInfo";
-import Lyrics from "@/src/models/Lyrics";
-import { genresInfo } from "@/src/utilities/constants";
-import Link from "next/link";
-import { useEffect, useRef, useState } from "react";
+export function generateMetadata({ params }: Params): Metadata {
+  const title = `${capitalize(
+    params.genre,
+    "-"
+  )} Lyrics | Midhah - Hamd, Naat, Manqbat and Durood o Salam lyrics platform`;
+  const description = `Explore your favorite ${params.genre} lyrics. Midhah مدحة is a leading & the most authentic lyrics searching platform for Hamd, Nasheed/Naat, Manqbat, and Durood o Salam. Download the app from Google Play Store.`;
+
+  return {
+    title,
+    description,
+    openGraph: {
+      title,
+      description,
+    },
+    twitter: {
+      title,
+      description,
+    },
+    alternates: {
+      canonical: `${WEB_BASE_URL}/${params.genre}`,
+    },
+  };
+}
 
 type Params = {
   params: { genre: string };
@@ -13,71 +33,7 @@ type Params = {
 export default function GenreListPage({ params }: Params) {
   const genre = params.genre;
 
-  const [lyrics, setLyrics] = useState<Lyrics[]>([]);
-  const [page, setPage] = useState(0);
-  const [isLoading, setIsLoading] = useState(true);
-  const [hasMoreData, setHasMoreData] = useState(true);
-
-  const [genreInfo, setGenreInfo] = useState<GenreInfo | null>(null);
-
-  const lastLyricRef = useRef<HTMLLIElement>(null);
-
-  useEffect(() => {
-    const options = {
-      root: null,
-      rootMargin: "20px",
-      threshold: 0,
-    };
-
-    const lastLyricNode = lastLyricRef.current;
-
-    const observer = new IntersectionObserver(([entry]) => {
-      if (entry.isIntersecting && !isLoading) {
-        setPage((prevPage) => prevPage + 1);
-      }
-    }, options);
-
-    if (lastLyricNode) {
-      observer.observe(lastLyricNode);
-    }
-
-    return () => {
-      if (lastLyricNode) {
-        observer.unobserve(lastLyricNode);
-      }
-    };
-  }, [isLoading, page, lastLyricRef]);
-
-  useEffect(() => {
-    setIsLoading(true);
-
-    if (genre && hasMoreData) {
-      setGenreInfo(getPageGenre(genre));
-
-      fetch(`https://api.midhah.com/v2/lyrics/${genre}?page=${page}&size=30`, {
-        method: "GET",
-        headers: {
-          "Content-Type": "application/json",
-        },
-      })
-        .then((response) => {
-          if (!response.ok) {
-            setHasMoreData(false);
-          }
-          return response.json();
-        })
-        .then((res) => {
-          if (res.data) {
-            setLyrics((prevLyrics) => [...prevLyrics, ...res.data]);
-          }
-          setIsLoading(false);
-        });
-    }
-  }, [genre, hasMoreData, page]);
-
-  function getPageGenre(genre: string) {
-    return genresInfo.filter((genreInfo) => genreInfo.path === genre)[0];
-  }
+  const genreInfo = getPageGenre(params.genre);
 
   return (
     <div className="container mx-auto w-full md:w-[85%]">
@@ -92,32 +48,7 @@ export default function GenreListPage({ params }: Params) {
         </div>
       </div>
 
-      <main className="flex min-h-[calc(100vh-575px)] flex-col items-center justify-center">
-        <ul className="w-full md:grid md:grid-cols-2">
-          {lyrics.map((lyric: Lyrics, index: number) => (
-            <Link href={`/${genre}/${lyric.slug}`} key={lyric.slug}>
-              <li
-                className="group relative my-1 flex flex-row hover:block"
-                ref={index === lyrics.length - 1 ? lastLyricRef : undefined}
-              >
-                <div className="flex flex-1 scale-100 cursor-pointer select-none items-center p-4 hover:bg-gray-50 group-hover:scale-0">
-                  <div className="mr-16 flex-1 pl-1">
-                    <h2 className="text-gray-600  ">{lyric.title}</h2>
-                    <h3 className="text-sm uppercase text-gray-400">
-                      {lyric.genre}
-                    </h3>
-                  </div>
-                </div>
-                <div className="poetry absolute top-1/2 -translate-y-1/2 scale-0 whitespace-pre-wrap text-center text-3xl group-hover:z-10 group-hover:w-full group-hover:scale-100 group-hover:bg-slate-50 group-hover:py-4 group-hover:transition-all">
-                  {lyric.preview}
-                </div>
-              </li>
-            </Link>
-          ))}
-        </ul>
-
-        {isLoading && <Loader />}
-      </main>
+      <RenderLyricsList genre={genre} />
     </div>
   );
 }
